feat: add document support for non-intrinsic code paths
