+++ mellea/stdlib/components/docs/document.py
@@ -6,10 +6,12 @@ typically attached to a ``Message`` via its ``documents`` parameter, enabling
 retrieval-augmented generation (RAG) workflows.
 """
 
+import collections.abc
+import warnings
+
 from ....core import CBlock, Component, ModelOutputThunk
 
 
-# TODO: Add support for passing in docs as model options.
 class Document(Component[str]):
     """A text passage with optional metadata for grounding model inputs.
 
@@ -56,3 +58,46 @@ class Document(Component[str]):
     def _parse(self, computed: ModelOutputThunk) -> str:
         """Parse the model output. Returns string value for now."""
         return computed.value if computed.value is not None else ""
+
+
+def _coerce_documents(
+    documents: collections.abc.Iterable[str | Document] | None,
+    *,
+    auto_doc_id: bool = False,
+) -> list[Document] | None:
+    """Convert an iterable of strings or Documents into a list of Documents.
+
+    Args:
+        documents: Strings, Document objects, a mix, or None.
+        auto_doc_id: When True, assign sequential string doc_id values
+            ("0", "1", ...) to Documents created from strings and warn about
+            existing Document objects that have no ``doc_id`` set.
+
+    Returns:
+        A list of Document objects, or None if the input was None.
+    """
+    if documents is None:
+        return None
+    result: list[Document] = []
+    for i, d in enumerate(documents):
+        if isinstance(d, str):
+            doc_id = str(i) if auto_doc_id else None
+            result.append(Document(text=d, doc_id=doc_id))
+        else:
+            if auto_doc_id and d.doc_id is None:
+                warnings.warn(
+                    f"Document at index {i} has no doc_id; results may omit "
+                    "document identification. Set doc_id on the Document or "
+                    "pass a plain string to auto-generate one.",
+                    UserWarning,
+                    stacklevel=3,
+                )
+            result.append(d)
+    return result
+
+
+def _coerce_document(document: str | Document) -> Document:
+    """Convert a single string or Document into a Document."""
+    if isinstance(document, str):
+        return Document(text=document)
+    return document
