+++ mellea/stdlib/components/intrinsic/rag.py
@@ -6,9 +6,8 @@ from ....backends.adapters import AdapterMixin
 from ...components import Document
 from ...context import ChatContext
 from ..chat import Message
+from ..docs.document import _coerce_document, _coerce_documents
 from ._util import (
-    _coerce_document,
-    _coerce_documents,
     _resolve_question,
     _resolve_response,
     call_intrinsic,
